+++ MarketPlace/Info.plist
@@ -2,6 +2,10 @@
 <!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
 <plist version="1.0">
 <dict>
+	<key>LSApplicationQueriesSchemes</key>
+	<array>
+		<string>kakaomap</string>
+	</array>
 	<key>KAKAO_API_TOKEN</key>
 	<string>$(KAKAO_API_TOKEN)</string>
 	<key>KAKAO_APP_KEY</key>
