+++ MarketPlace/LocationManager.swift
@@ -2,7 +2,7 @@ import Foundation
 import CoreLocation
 import MapKit
 
-class LocationManager: NSObject, ObservableObject, CLLocationManagerDelegate {
+final class LocationManager: NSObject, ObservableObject, CLLocationManagerDelegate {
     static let shared = LocationManager()
     private var manager: CLLocationManager = CLLocationManager()
     
@@ -48,30 +48,5 @@ class LocationManager: NSObject, ObservableObject, CLLocationManagerDelegate {
     func locationManager(_ manager: CLLocationManager, didFailWithError error: Error) {
         // Handle errors when the location manager fails to get the user's location
         print("Location manager failed with error: \(error.localizedDescription)")
-    }
-}
-
-enum AddressError: Error {
-    case failedToConvertAddress
-}
-
-final class ConvertAddress {
-    /// - NOTE: 좌표 -> 도로명 주소
-    func loadCurrentUserRoadAddress(latitude: Double, longitude: Double) async throws -> String {
-        let geoCoder = CLGeocoder()
-        let places = try await geoCoder.reverseGeocodeLocation(CLLocation(latitude: latitude, longitude: longitude))
-        guard let place = places.last,
-              let sido = place.administrativeArea,
-              let gugun = place.locality else { throw AddressError.failedToConvertAddress }
-        return "\(sido) \(gugun)"
-    }
-    
-    /// - NOTE: 도로명 주소 -> 좌표
-    func getCoordinateFromRoadAddress(from address: String) async throws -> CLLocationCoordinate2D {
-        let geoCoder = CLGeocoder()
-        let places = try await geoCoder.geocodeAddressString(address)
-        guard let place = places.last,
-              let coordinate = place.location?.coordinate else { throw AddressError.failedToConvertAddress }
-        return coordinate
     }
 }
