+++ MarketPlace/Model/Market/MarketDetailModel.swift
@@ -1,5 +1,6 @@
-import Foundation
 
+import Foundation
+import CoreLocation
 
 struct MarketDetailModel: Codable {
     let marketId: Int
@@ -10,6 +11,12 @@ struct MarketDetailModel: Codable {
     let phoneNumber: String
     let address: String
     let imageResList: [ImageResource]
+    
+    var position: CLLocationCoordinate2D?
+    
+    enum CodingKeys: String, CodingKey {
+        case marketId, name, description, operationHours, closedDays, phoneNumber, address, imageResList
+    }
 }
 
 struct ImageResource: Codable {
